EPMDJ-2221 FunctionInfo.h is renamed to CorDataStructures.h and will define info structures obtained via COM interfaces.
