+++ Drill4dotNet/Drill4dotNet/CorProfilerInfo.h
@@ -5,7 +5,7 @@
 
 #include "framework.h"
 #include "OutputUtils.h"
-#include "FunctionInfo.h"
+#include "CorDataStructures.h"
 #include "ComWrapperBase.h"
 #include "MetaDataImport.h"
